requirements.txt
asgiref~=3.8.1
channels~=4.1.0
daphne~=4.1.0
Django~=5.0.3
django-cors-headers~=4.3.1
djangorestframework~=3.15.1
djangorestframework-simplejwt~=5.3.1
drf-yasg~=1.21.7
pillow~=10.3.0
psycopg2-binary~=2.9.9
python-dotenv~=1.0.1
PyJWT~=2.8.0